requirements.txt
openai-agents[voice]
aiofiles
elevenlabs
openai
pydantic
pydub
python-dotenv
google-genai
